refactor(fw): refactor vm/bytecode
